Javadoc and Linter
